Bug fixes and screens widget extraction
