+++ pchaa/pchaa_flutter/lib/screens/available_menu_items_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pchaa_client/pchaa_client.dart';
-import 'package:pchaa_flutter/utils/url_utils.dart';
 import '../services/app_services.dart';
+import '../widgets/menu_items/available_menu_item_card.dart';
 import 'menu_item_detail_screen.dart';
 
 class AvailableMenuItemsScreen extends StatefulWidget {
@@ -117,149 +117,19 @@ class _AvailableMenuItemsScreenState extends State<AvailableMenuItemsScreen> {
         padding: const EdgeInsets.all(16),
         itemCount: _menuItems!.length,
         itemBuilder: (context, index) {
-          return _buildMenuItemCard(_menuItems![index]);
-        },
-      ),
-    );
-  }
-
-  Widget _buildMenuItemCard(AvailableMenuItem menuItem) {
-    return Card(
-      margin: const EdgeInsets.only(bottom: 16),
-      elevation: 4,
-      clipBehavior: Clip.antiAlias,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: InkWell(
-        onTap: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => MenuItemDetailScreen(menuItem: menuItem),
-            ),
+          final menuItem = _menuItems![index];
+          return AvailableMenuItemCard(
+            menuItem: menuItem,
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => MenuItemDetailScreen(menuItem: menuItem),
+                ),
+              );
+            },
           );
         },
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // Image section with local greyscale logic
-            if (menuItem.imageUrl != null && menuItem.imageUrl!.isNotEmpty)
-              ColorFiltered(
-                // Only greyscale the image if not for sale
-                colorFilter: menuItem.forSale
-                    ? const ColorFilter.mode(
-                        Colors.transparent,
-                        BlendMode.multiply,
-                      )
-                    : const ColorFilter.matrix(<double>[
-                        0.2126,
-                        0.7152,
-                        0.0722,
-                        0,
-                        0,
-                        0.2126,
-                        0.7152,
-                        0.0722,
-                        0,
-                        0,
-                        0.2126,
-                        0.7152,
-                        0.0722,
-                        0,
-                        0,
-                        0,
-                        0,
-                        0,
-                        1,
-                        0,
-                      ]),
-                child: AspectRatio(
-                  aspectRatio: 16 / 9,
-                  child: Image.network(
-                    UrlUtils.getDisplayableImageUrl(menuItem.imageUrl),
-                    fit: BoxFit.cover,
-                    errorBuilder: (context, error, stackTrace) => Container(
-                      color: Colors.grey[200],
-                      child: const Icon(
-                        Icons.broken_image,
-                        size: 48,
-                        color: Colors.grey,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-
-            // Content section (Always in color)
-            Padding(
-              padding: const EdgeInsets.all(16),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Expanded(
-                        child: Text(
-                          menuItem.name,
-                          style: Theme.of(context).textTheme.titleLarge
-                              ?.copyWith(
-                                fontWeight: FontWeight.bold,
-                              ),
-                        ),
-                      ),
-                      _buildStatusBadge(menuItem.forSale),
-                    ],
-                  ),
-                  const SizedBox(height: 12),
-                  Row(
-                    children: [
-                      Text(
-                        // Removed the $ sign here
-                        menuItem.basePrice.toStringAsFixed(2),
-                        style: const TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.bold,
-                          color: Colors.green,
-                        ),
-                      ),
-                      const SizedBox(width: 16),
-                      const Icon(
-                        Icons.timer_outlined,
-                        size: 18,
-                        color: Colors.orange,
-                      ),
-                      const SizedBox(width: 4),
-                      Text('${menuItem.timeToPrepare} min'),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildStatusBadge(bool forSale) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-      decoration: BoxDecoration(
-        color: forSale
-            ? Colors.green.withOpacity(0.1)
-            : Colors.red.withOpacity(0.1),
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: forSale ? Colors.green : Colors.red),
-      ),
-      child: Text(
-        forSale ? 'AVAILABLE' : 'UNAVAILABLE',
-        style: TextStyle(
-          fontSize: 10,
-          fontWeight: FontWeight.bold,
-          color: forSale ? Colors.green[700] : Colors.red[700],
-        ),
       ),
     );
   }
