+++ pchaa/pchaa_flutter/lib/screens/menu_detail_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:pchaa_client/pchaa_client.dart';
 import 'package:pchaa_flutter/services/app_services.dart';
-import 'package:pchaa_flutter/utils/url_utils.dart';
-import 'package:pchaa_flutter/widgets/customization_choices.dart';
+import 'package:pchaa_flutter/widgets/menu_detail/menu_detail_bottom_bar.dart';
+import 'package:pchaa_flutter/widgets/menu_detail/menu_detail_content.dart';
 
 class MenuDetailScreen extends StatefulWidget {
   final int menuId;
@@ -13,7 +13,7 @@ class MenuDetailScreen extends StatefulWidget {
 }
 
 class _MenuDetailScreenState extends State<MenuDetailScreen> {
-  bool _isShopOpen = isShopOpen;
+  final bool _isShopOpen = isShopOpen;
   AvailableMenuItem? _menuDetail;
   bool _isLoading = false;
   String? _errorMsg;
@@ -22,7 +22,6 @@ class _MenuDetailScreenState extends State<MenuDetailScreen> {
   final Map<String, List<SelectedOption>> _selectedOptionsByGroup = {};
   double _addOnTotal = 0;
 
-  static const _imageHeight = 200.0;
   static const _primaryColor = Color(0xFF5A9CB5);
 
   @override
@@ -132,14 +131,16 @@ class _MenuDetailScreenState extends State<MenuDetailScreen> {
         _quantity,
         _messageController.text.trim(),
       );
+      if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(
           content: Text("เพิ่มเข้าตะกร้าแล้ว"),
         ),
       );
-      Navigator.pop(context);
+      if (mounted) Navigator.pop(context);
     } catch (e) {
       debugPrint(e.toString());
+      if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
           content: Text("Error: ${e.toString()}"),
@@ -147,10 +148,11 @@ class _MenuDetailScreenState extends State<MenuDetailScreen> {
         ),
       );
     } finally {
-      if (!mounted) return;
-      setState(() {
-        _isLoading = false;
-      });
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+        });
+      }
     }
   }
 
@@ -160,9 +162,24 @@ class _MenuDetailScreenState extends State<MenuDetailScreen> {
       children: [
         Scaffold(
           body: SafeArea(
-            child: _buildContent(_menuDetail),
+            child: MenuDetailContent(
+      menuDetail: _menuDetail,
+      errorMsg: _errorMsg,
+      messageController: _messageController,
+      getSelectedNamesForGroup: _selectedNamesForGroup,
+      onSingleChanged: (group, value) {
+        setState(() {
+          _setSingleSelection(group, value);
+        });
+      },
+      onMultiChanged: (group, values) {
+        setState(() {
+          _setMultiSelection(group, values);
+        });
+      },
+    ),
           ),
-          bottomNavigationBar: BottomBar(
+          bottomNavigationBar: MenuDetailBottomBar(
             isLogin: googleAuthService.isLoggedIn,
             isShopOpened: _isShopOpen,
             quantity: _quantity,
@@ -190,231 +207,6 @@ class _MenuDetailScreenState extends State<MenuDetailScreen> {
             ),
           ),
       ],
-    );
-  }
-
-  Widget _buildContent(AvailableMenuItem? menuDetail) {
-    if (_errorMsg != null) {
-      return Center(child: Text(_errorMsg!));
-    }
-    if (menuDetail == null) {
-      return const Center(child: Text("No Menu Detail"));
-    }
-
-    return SingleChildScrollView(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          _buildHeader(context),
-          Padding(
-            padding: const EdgeInsets.only(
-              left: 30,
-              top: 10,
-              right: 30,
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  menuDetail.name,
-                  style: const TextStyle(
-                    fontSize: 25,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                const Padding(
-                  padding: EdgeInsets.symmetric(vertical: 4),
-                  child: Divider(),
-                ),
-
-                if (menuDetail.customization.isNotEmpty)
-                  ...menuDetail.customization.map((customization) {
-                    final selectedValues = _selectedNamesForGroup(
-                      customization,
-                    );
-                    final selectedValue = selectedValues.isNotEmpty
-                        ? selectedValues.first
-                        : null;
-                    return CustomizationChoices(
-                      customizationGroup: customization,
-                      selectedValue: selectedValue,
-                      selectedValues: selectedValues,
-                      onSingleChanged: (value) {
-                        setState(() {
-                          _setSingleSelection(customization, value);
-                        });
-                      },
-                      onMultiChanged: (values) {
-                        setState(() {
-                          _setMultiSelection(customization, values);
-                        });
-                      },
-                    );
-                  }),
-                const Padding(
-                  padding: EdgeInsets.symmetric(vertical: 4),
-                  child: Divider(),
-                ),
-                Text(
-                  "เพิ่มเติม",
-                  style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-                ),
-                SizedBox(
-                  height: 10,
-                ),
-                TextField(
-                  controller: _messageController,
-                  maxLines: null,
-                  keyboardType: TextInputType.multiline,
-                  decoration: InputDecoration(
-                    hintText: 'โน๊ตเพิ่มเติม',
-                    border: OutlineInputBorder(),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildHeader(BuildContext context) {
-    final imageUrl = UrlUtils.getDisplayableImageUrl(_menuDetail?.imageUrl);
-    final ImageProvider imageProvider = imageUrl.isEmpty
-        ? const AssetImage("assets/images/food.jpg")
-        : NetworkImage(imageUrl);
-    return Stack(
-      children: [
-        Container(
-          height: _imageHeight,
-          decoration: BoxDecoration(
-            image: DecorationImage(
-              image: imageProvider,
-              fit: BoxFit.cover,
-            ),
-          ),
-        ),
-        Container(
-          height: _imageHeight,
-          decoration: BoxDecoration(
-            gradient: LinearGradient(
-              begin: Alignment.topCenter,
-              end: Alignment.bottomCenter,
-              colors: [
-                Colors.black.withValues(alpha: 0.3),
-                Colors.black.withValues(alpha: 0.5),
-              ],
-            ),
-            color: Colors.white,
-          ),
-        ),
-        Positioned(
-          top: 30,
-          left: 10,
-          child: IconButton(
-            onPressed: () {
-              Navigator.pop(context);
-            },
-            icon: const Icon(
-              Icons.chevron_left,
-              color: Colors.black,
-              size: 30,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class BottomBar extends StatelessWidget {
-  final int quantity;
-  final bool isLogin;
-  final AvailableMenuItem? menuDetail;
-  final bool isShopOpened;
-  final ValueChanged<int> onQuantityChanged;
-  final VoidCallback? onAddToCart;
-  final double basePrice;
-  final double addOnTotal;
-  final Color primaryColor;
-  const BottomBar({
-    super.key,
-    required this.isLogin,
-    required this.isShopOpened,
-    required this.quantity,
-    required this.menuDetail,
-    required this.onQuantityChanged,
-    required this.basePrice,
-    required this.addOnTotal,
-    required this.primaryColor,
-    this.onAddToCart,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final total = (basePrice + addOnTotal) * quantity;
-    return Container(
-      decoration: BoxDecoration(
-        border: Border(
-          top: BorderSide(
-            color: Colors.grey.withValues(alpha: 1),
-            width: 0.5,
-          ),
-        ),
-      ),
-      margin: const EdgeInsets.only(bottom: 30),
-      padding: const EdgeInsets.only(left: 10, right: 10, top: 10),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Row(
-            children: [
-              IconButton(
-                onPressed: quantity > 1 && isLogin
-                    ? () => onQuantityChanged(quantity - 1)
-                    : null,
-                icon: const Icon(Icons.remove),
-                style: const ButtonStyle(),
-              ),
-              Text(
-                quantity.toString(),
-                style: const TextStyle(fontSize: 20),
-              ),
-              IconButton(
-                onPressed: isLogin ? () => onQuantityChanged(quantity + 1) : null,
-                icon: const Icon(Icons.add),
-                style: const ButtonStyle(),
-              ),
-            ],
-          ),
-          Expanded(
-            child: ElevatedButton(
-              onPressed: isShopOpened && quantity > 0 && menuDetail?.forSale == true && isLogin ? onAddToCart : null,
-              style: ElevatedButton.styleFrom(
-                backgroundColor: primaryColor,
-                foregroundColor: Colors.black,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(10),
-                ),
-              ),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(
-                    "Add To Cart",
-                    style: const TextStyle(fontWeight: FontWeight.bold),
-                  ),
-                  Text(
-                    total.toStringAsFixed(2),
-                    style: const TextStyle(fontWeight: FontWeight.bold),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
