+++ pchaa/pchaa_flutter/lib/widgets/common/google_login_button.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:pchaa_client/pchaa_client.dart';
-import '../services/app_services.dart';
-import '../services/cart_service.dart';
+import '../../services/app_services.dart';
+import '../../services/cart_service.dart';
 
 class GoogleLoginButton extends StatefulWidget {
   final Function()? onLoginSuccess;
