+++ pchaa/pchaa_flutter/lib/widgets/menu_screen/menu_screen_header.dart
@@ -1,0 +1,93 @@
+import 'package:flutter/material.dart';
+
+class MenuScreenHeader extends StatelessWidget {
+  final VoidCallback onBack;
+
+  const MenuScreenHeader({
+    super.key,
+    required this.onBack,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        // Background food image
+        Container(
+          height: 300,
+          decoration: const BoxDecoration(
+            image: DecorationImage(
+              image: AssetImage("assets/images/food.jpg"),
+              fit: BoxFit.cover,
+            ),
+          ),
+        ),
+
+        // Gradient / overlay container
+        Container(
+          height: 300,
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topCenter,
+              end: Alignment.bottomCenter,
+              colors: [
+                Colors.black.withValues(alpha: 0.3),
+                Colors.black.withValues(alpha: 0.5),
+              ],
+            ),
+            color: Colors.white,
+          ),
+        ),
+        Positioned(
+          top: 30,
+          left: 10,
+          child: IconButton(
+            onPressed: onBack,
+            icon: const Icon(
+              Icons.chevron_left,
+              color: Colors.black,
+              size: 30,
+            ),
+          ),
+        ),
+
+        // Positioned welcome text and button
+        const Positioned(
+          bottom: 10,
+          left: 30,
+          right: 20,
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            crossAxisAlignment: CrossAxisAlignment.end,
+            children: [
+              Flexible(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      "MENU",
+                      overflow: TextOverflow.ellipsis,
+                      maxLines: 1,
+                      style: TextStyle(
+                        color: Colors.white,
+                        fontSize: 30,
+                        fontWeight: FontWeight.bold,
+                        shadows: [
+                          Shadow(
+                            offset: Offset(2, 1),
+                            blurRadius: 10,
+                            color: Colors.grey,
+                          ),
+                        ],
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
