+++ pchaa/pchaa_flutter/lib/widgets/menu_screen/menu_search_bar.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class MenuSearchBar extends StatelessWidget {
+  final TextEditingController controller;
+  final ValueChanged<String> onChanged;
+
+  const MenuSearchBar({
+    super.key,
+    required this.controller,
+    required this.onChanged,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(top: 10, left: 20, right: 20),
+      padding: const EdgeInsets.symmetric(horizontal: 12),
+      height: 44,
+      decoration: BoxDecoration(
+        color: const Color.fromARGB(87, 104, 159, 180),
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Row(
+        children: [
+          const Icon(Icons.search, color: Colors.grey),
+          const SizedBox(width: 8),
+          Expanded(
+            child: TextField(
+              controller: controller,
+              onChanged: onChanged,
+              decoration: const InputDecoration(
+                hintText: 'Search',
+                border: InputBorder.none,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
