+++ pchaa/pchaa_flutter/lib/widgets/owner_page/owner_page_header.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pchaa_flutter/services/app_services.dart';
-import 'package:pchaa_flutter/widgets/google_login_button.dart';
+import 'package:pchaa_flutter/widgets/common/google_login_button.dart';
 
 class OwnerPageHeader extends StatelessWidget {
   final VoidCallback onLoginSuccess;
