+++ pchaa/pchaa_server/lib/src/models/cart_detail.yaml
@@ -2,3 +2,4 @@ class: CartDetail
 fields:
   cart: Cart
   menuItemName: String
+  imageUrl: String
